--- src/main/java/edu/kis/vh/nursery/list/Node.java
@@ -1,19 +1,0 @@
-package edu.kis.vh.nursery.list;
-
-public class Node {
-
-	private int value;
-	public Node prev, next;
-
-	public Node(int i) {
-		setValue(i);
-	}
-
-	public int getValue() {
-		return value;
-	}
-
-	public void setValue(int value) {
-		this.value = value;
-	}
-}
